Added swagger, and minor code refactoring

+++ src/main/java/com/nikolanedeljkovic/flightadvisor/service/CityServiceImpl.java
@@ -6,7 +6,9 @@ import java.util.Optional;
 
 import javax.transaction.Transactional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.nikolanedeljkovic.flightadvisor.domain.city.City;
 import com.nikolanedeljkovic.flightadvisor.domain.city.Comment;
@@ -24,47 +26,51 @@ public class CityServiceImpl implements CityService {
 	private final CityRepository cityRepository;
 	private final CommentRepository commentRepository;
 	private final UserRepository userRepository;
-	
+
 	@Override
 	public City addCity(City city) {
 		return cityRepository.save(city);
 	}
-	
+
 	@Override
 	public List<City> getAllCities() {
 		return cityRepository.findAll();
 	}
-	
+
 	@Override
 	public List<City> getAllCities(Optional<Integer> numberOfComments) {
 		return cityRepository.getAllCitiesWithSpecifiedNumberOfComments(numberOfComments);
 	}
-	
+
 	@Override
 	public Comment postComment(String cityName, Comment comment, String username) {
-		return commentRepository.save(Comment.builder()
-				.city(cityRepository.findByName(cityName))
-				.user(userRepository.findByUsername(username))
-				.comment(comment.getComment())
-				.createdAt(LocalDateTime.now())
-				.modifiedAt(LocalDateTime.now())
-				.build());
+		City city = cityRepository.findByName(cityName);
+		if (city != null) {
+			return commentRepository.save(Comment.builder().city(city).user(userRepository.findByUsername(username))
+					.comment(comment.getComment()).createdAt(LocalDateTime.now()).modifiedAt(LocalDateTime.now())
+					.build());
+		} throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Unable to post comment. Cityname not valid.");
 	}
 
 	@Override
-	public String deleteComment(Long commentId) {
-		commentRepository.deleteById(commentId);
-		return "Deleted";
+	public String deleteComment(Long commentId, String username) {
+		if (commentRepository.findById(commentId).get().getUser().getUsername().equals(username)) {
+			commentRepository.deleteById(commentId);
+			return "Deleted";
+		}
+		throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can't delete comment posted by someone else.");
 	}
 
 	@Override
-	public Comment updateComment(Long commentId, String comment) {
+	public Comment updateComment(Long commentId, String comment, String username) {
 		Optional<Comment> oldComment = commentRepository.findById(commentId);
-		Comment newComment = oldComment.get();
-		newComment.setComment(comment);
-		newComment.setModifiedAt(LocalDateTime.now());
-		return commentRepository.save(newComment);
+		if (oldComment.isPresent() && oldComment.get().getUser().getUsername().equals(username)) {
+			Comment newComment = oldComment.get();
+			newComment.setComment(comment);
+			newComment.setModifiedAt(LocalDateTime.now());
+			return commentRepository.save(newComment);
+		}
+		throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can't update comment posted by someone else.");
 	}
-	
 
 }
